refactor(console): move home child components to `components ` folder
