+++ gravitee-apim-console-webui/src/management/home/components/gio-api-lifecycle-state/gio-api-lifecycle-state.component.html
