+++ gravitee-apim-console-webui/src/management/home/components/gio-api-state/gio-api-state.component.html
