+++ gravitee-apim-console-webui/src/management/home/components/gio-quick-time-range/gio-quick-time-range.component.html
