+++ gravitee-apim-console-webui/src/management/home/components/gio-request-stats/gio-request-stats.component.scss
