+++ gravitee-apim-console-webui/src/management/home/components/gio-top-apis-table/gio-top-apis-table.component.html
